Extract findPositionAtPixel from paginatedPixelToPosition
Pure extraction. The layout-shaped portion of
paginatedPixelToPosition (find line, find run, find char) moves
into a new find-position-at-pixel.ts that operates on any
DocumentLayout in layout-local coordinates. The paginated wrapper
keeps the page-finding step and translates page-local pixels into
layout-local before delegating.

No behaviour change for the docs editor. Phase 5a slides text-box
editor consumes findPositionAtPixel for cursor / selection
hit-testing inside a single text-box.

Refs packages/slides/spike/docs-richtext-audit.md 'Selection / cursor'
and 'Required exports'.
